refactor(telemetry): move injected-script enrichment to a web-only module
