+++ cspell.json
@@ -527,6 +527,8 @@
         "colorscale",
         "commentbubbles",
         "contenteditable",
+        "conv",
+        "convertexperiments",
         "copiloted",
         "copiloting",
         "copyable",
@@ -639,6 +641,7 @@
         "getsentry",
         "gibsdk",
         "glcode",
+        "googletagmanager",
         "googleusercontent",
         "gorhom",
         "gpgsign",
@@ -657,6 +660,7 @@
         "helpsite",
         "hexcode",
         "hibob",
+        "hotjar",
         "hrefs",
         "hris",
         "hybridapp",
@@ -686,6 +690,7 @@
         "jridgewell",
         "jsSrcsDir",
         "jsbundle",
+        "ketchcdn",
         "keyalg",
         "keycap",
         "keycommand",
@@ -801,6 +806,7 @@
         "onxy",
         "oxfmt",
         "openxmlformats",
+        "optimizely",
         "ordinality",
         "organisation",
         "originalamount",
